Architecture fix and use cases refactor

+++ main/src/main/java/com/picpay/desafio/android/main/presentation/dataui/UserDataUi.kt
@@ -5,8 +5,8 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class UserDataUi(
-    val id: Int,
+    val id: Long,
     val img: String,
     val name: String,
     val userName: String
-): Parcelable
+) : Parcelable
